% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_heart_disease.R
\name{predict_heart_disease}
\alias{predict_heart_disease}
\title{Predict Heart Disease Risk}
\usage{
predict_heart_disease(new_data, type = "class")
}
\arguments{
\item{new_data}{A data frame containing clinical features.}

\item{type}{Character. "class" for category or "prob" for probability.}
}
\value{
A vector or matrix of predictions.
}
\description{
This function utilizes a pre-trained Random Forest model to predict
heart disease risk. Developed by Sirui Cheng.
}
\examples{
\dontrun{
# Example usage:
library(Group3)
# results <- predict_heart_disease(your_data)
}
}
